+++ pid_client/control/pid.h
@@ -14,7 +14,7 @@ public:
 	double umax;
 	bool inverted;
 	PIDParameters();
-	PIDParameters(double,double,double,double,int,double,double,double,bool);
+	PIDParameters(double,double,double,double,double,double,double,double,bool);
 
 	friend std::ostream& operator <<(std::ostream&, const PIDParameters&);
 };
