+++ pid_client/main.cc
@@ -56,7 +56,7 @@ bool Parameters::update_parameters()
 			     pt.get<double>("PID.Ti"),
 			     pt.get<double>("PID.Td"),
 			     pt.get<double>("PID.Tr"),
-				 period,
+				 (double) period,
 			     pt.get<double>("PID.ref"),
 			     pt.get<double>("PID.umin"),
 			     pt.get<double>("PID.umax"),
